+++ tests/testthat/test-Splits.R
@@ -6,6 +6,9 @@ expect_splits_equal <- function(s1, s2, ...) {
 test_that("as.Splits()", {
   A <- FALSE
   B <- TRUE
+  expect_equal(as.Splits(BalancedTree(0)),
+               NULL
+               )
   expect_equal(strsplit(capture_output(summary(as.Splits(c(A, A, B, B)))),
                         "\n")[[1]],
                c("1 bipartition split dividing 4 tips, t1 .. t4",